fix: Send Proposal hook added
